+++ lib/shopify_cli/theme/extension/host_theme.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require_relative "../syncer"
 require_relative "../development_theme"
+require_relative 'host_theme/host_theme_progress_bar'
 require "shopify_cli/git"
 require "tmpdir"
 require "fileutils"
@@ -8,34 +9,6 @@ require "fileutils"
 module ShopifyCLI
   module Theme
     module Extension
-      # TODO: MOVEEEEEEE
-      class HostThemeProgressBar
-        GIT_CLONE_PROGRESS_SHARE = 0.2
-        SYNC_PROGRESS_SHARE = 0.8
-
-        def initialize(syncer, dir)
-          # TODO: fix sig
-          @syncer = syncer
-          @dir = dir
-        end
-
-        def progress(method, **args)
-          @syncer.lock_io!
-          CLI::UI::Progress.progress do |bar|
-            Git.public_send(:clone, "https://github.com/Shopify/dawn.git", @dir) do |percent|
-              bar.tick(set_percent: percent * GIT_CLONE_PROGRESS_SHARE)
-            end
-
-            @syncer.public_send(method, **args) do |left, total|
-              bar.tick(set_percent: (1 - left.to_f / total) * SYNC_PROGRESS_SHARE + GIT_CLONE_PROGRESS_SHARE)
-            end
-
-            bar.tick(set_percent: 1)
-          end
-          @syncer.unlock_io!
-        end
-      end
-
       class HostTheme < DevelopmentTheme
         def id
           ShopifyCLI::DB.get(:host_theme_id)
